mmap testing and bug fix

+++ base/mmap.c
@@ -19,16 +19,17 @@
 
 size_t getpagesize();
 
-void *mmap_allocate(unsigned int size) {
+void *mmap_allocate(unsigned long long size) {
+	/*println("size=%llu,as=%llu", size, mmap_aligned_size(size));*/
 	return mmap(NULL, mmap_aligned_size(size), PROT_READ | PROT_WRITE,
 				MAP_PRIVATE | MAP_ANONYMOUS, -1, 0);
 }
 
-void mmap_free(void *ptr, unsigned int size) {
+void mmap_free(void *ptr, unsigned long long size) {
 	if (munmap(ptr, mmap_aligned_size(size))) panic("unexpected mmap error!");
 }
 
-unsigned int mmap_aligned_size(unsigned int size) {
-	unsigned int page_size = getpagesize();
+unsigned long long mmap_aligned_size(unsigned long long size) {
+	unsigned long long page_size = getpagesize();
 	return (size + (page_size - 1)) & ~(page_size - 1);
 }
